# docs/swagger.yaml
definitions:
...
  models.Template:
    properties:
      created_at:
        type: string
      description:
        description: Optional description
        type: string
      event_type:
        description: Event type (e.g., "push", "pull_request")
        type: string
      id:
        type: integer
      project_id:
        description: Foreign key to Project
        type: string
      template:
        description: Template content
        type: string
      updated_at:
        type: string
    type: object
  utils.Response:
    properties:
      data:
        type: object
      error:
        items:
          type: string
        type: array
      status:
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /api/v1/projects/{projectID}/templates:
    get:
      description: Retrieve all templates associated with a specific project
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Template'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Get all templates for a project
      tags:
      - Templates
    post:
      description: Create a new template for a project
      parameters:
      - description: Template data
        in: body
        name: template
        required: true
        schema:
          $ref: '#/definitions/models.Template'
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Template'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Add a new template
      tags:
      - Templates
  /api/v1/projects/{projectID}/templates/event-type/{eventType}:
    get:
      description: Retrieve a template associated with a specific project and event
        type
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Event type
        in: path
        name: eventType
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Template'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Get a template by event type
      tags:
      - Templates
  /api/v1/templates/{templateID}:
    delete:
      description: Delete a template by its ID
      parameters:
      - description: Template ID
        in: path
        name: templateID
        required: true
        type: string
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Delete a template
      tags:
      - Templates
    put:
      description: Update the details of an existing template
      parameters:
      - description: Template ID
        in: path
        name: templateID
        required: true
        type: string
      - description: Updated template data
        in: body
        name: template
        required: true
        schema:
          $ref: '#/definitions/models.Template'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Template'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Update an existing template
      tags:
      - Templates
...
swagger: "2.0"